fix: resolve multiple UI bugs
